Cleanup flow tools. Closes 2392 and 2350.  - fixed context labels for segments  - removed unused util methods  - misc flowtoolutil cleanup  - fixed cyclic flows error  - fixed showing ETE Flow segments that are ETE Flows
